some changes done to make the application maintainable & payslip generation api is done

+++ Backend/HRManagementSystem/HRManagementSystem/src/main/java/com/vmpkp/HRManagementSystem/Models/Position.java
@@ -20,7 +20,7 @@ public class Position {
 
     private String PositionTitle;
 
-    private int salary;
+    private Double salary;
 
     private String JobDescription;
